[TP-1037] Refactored out terminator to its own file. Added version details struct and some code to pull the canonical version from an S3 bucket.
